ADDED: Exception Handler And List Added In Another Table

+++ src/main/java/com/bridgelabz/greetingapp/service/EmployeePayrollService.java
@@ -1,12 +1,12 @@
 package com.bridgelabz.greetingapp.service;
 
 import com.bridgelabz.greetingapp.DTO.EmployeeDTO;
+import com.bridgelabz.greetingapp.exception.CustomException;
 import com.bridgelabz.greetingapp.model.EmployeeData;
 import com.bridgelabz.greetingapp.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -19,24 +19,50 @@ public class EmployeePayrollService implements IEmployeePayrollService {
         return employeeRepository.save(employeeData);
     }
     public Optional<EmployeeData> findEmployeeID(int id){
-        return employeeRepository.findById(id);
+        if(employeeRepository.findById(id).isPresent()){
+            return employeeRepository.findById(id);
+        }else {
+            throw new CustomException("Employee with "+id+" id is not found");
+        }
     }
+    /**
+     * Used try catch block because employeeRepository.findById(id).get() was giving NoSuchElementException
+     * exception when it was unable to find the ID in the Database
+     * */
+//    public EmployeeData updateEmployeeID(EmployeeDTO employeeDTO, int id){
+//        try{
+//            EmployeeData employeeData = employeeRepository.findById(id).get();
+//            employeeData.setFirstName(employeeDTO.getFirstName());
+//            employeeData.setSalary(employeeDTO.getSalary());
+//            return employeeRepository.save(employeeData);
+//
+//        }catch (NoSuchElementException e){
+//            return null;
+//        }
+//    }
     public EmployeeData updateEmployeeID(EmployeeDTO employeeDTO, int id){
-        try{
-            EmployeeData employeeData = employeeRepository.findById(id).get() ;
+        if(employeeRepository.findById(id).isPresent()) {
+            EmployeeData employeeData = employeeRepository.findById(id).get();
             employeeData.setFirstName(employeeDTO.getFirstName());
             employeeData.setSalary(employeeDTO.getSalary());
             return employeeRepository.save(employeeData);
-
-        }catch (NoSuchElementException e){
-            return null;
+        }else{
+            throw new CustomException("Employee with "+id+" id is not found");
         }
     }
     public Integer deleteEmployeeID(int id) {
-        employeeRepository.deleteById(id);
-        return id;
+        if (employeeRepository.findById(id).isPresent()) {
+            employeeRepository.deleteById(id);
+            return id;
+        }else{
+            throw new CustomException("Employee with "+id+" id is not found");
+        }
     }
     public List<EmployeeData> findAllGreets(){
-        return employeeRepository.findAll();
+        if(!employeeRepository.findAll().isEmpty()){
+            return employeeRepository.findAll();
+        }else {
+            throw new CustomException("Table is Empty");
+        }
     }
 }
